This is the initial commit to the project
This commit contains the initial release of the middleware that has just
enough functionality to be meaningfully functional. What it doesn't do,
that I can think of as I type this, is provide a factory class to
instantiate the middleware and set the old and new domain (retrieving
them from an app's config, environment variables, etc), or a
ConfigProvider class for when used with Mezzio.